Multi is an object for grouping multiple DB operations.
Multi makes it possible to pack operations that should be performed in a single database transaction and gives a way to introspect the queued operations without actually performing them. Each operation is given a name that is unique and will identify its result in case of success or failure.

All operations will be executed in the order they were added.

The Multi structure should be considered opaque, accessing fields or directly modifying them is not advised.